+++ internal/ingress/controller/kong.go
@@ -47,26 +47,39 @@ func (n *KongController) OnUpdate(state *parser.KongState) error {
 		return err
 	}
 
-	jsonConfig, err := json.Marshal(targetContent)
-	if err != nil {
-		return errors.Wrap(err,
-			"marshaling Kong declarative configuration to JSON")
-	}
-	shaSum := sha256.Sum256(jsonConfig)
-	if reflect.DeepEqual(n.runningConfigHash, shaSum) {
-		glog.Info("no configuration change, skipping sync to Kong")
-		return nil
+	var shaSum []byte
+	// disable optimization if reverse sync is enabled
+	if !n.cfg.EnableReverseSync {
+		shaSum, err = generateSHA(targetContent)
+		if err != nil {
+			return err
+		}
+		if reflect.DeepEqual(n.runningConfigHash, shaSum) {
+			glog.Info("no configuration change, skipping sync to Kong")
+			return nil
+		}
 	}
 	if n.cfg.InMemory {
 		err = n.onUpdateInMemoryMode(targetContent)
 	} else {
 		err = n.onUpdateDBMode(targetContent)
 	}
-	if err == nil {
-		glog.Info("successfully synced configuration to Kong")
-		n.runningConfigHash = shaSum
+	if err != nil {
+		return err
 	}
-	return err
+	n.runningConfigHash = shaSum
+	glog.Info("successfully synced configuration to Kong")
+	return nil
+}
+
+func generateSHA(targetContent *file.Content) ([]byte, error) {
+	jsonConfig, err := json.Marshal(targetContent)
+	if err != nil {
+		return nil, errors.Wrap(err,
+			"marshaling Kong declarative configuration to JSON")
+	}
+	shaSum := sha256.Sum256(jsonConfig)
+	return shaSum[:], nil
 }
 
 func cleanUpNullsInPluginConfigs(state *file.Content) {
@@ -278,7 +291,7 @@ func (n *KongController) toDeckContent(
 	})
 
 	for _, c := range k8sState.Certificates {
-		cert := file.FCertificate{Certificate: c.Certificate}
+		cert := getFCertificateFromKongCert(c.Certificate)
 		content.Certificates = append(content.Certificates, cert)
 	}
 	sort.SliceStable(content.Certificates, func(i, j int) bool {
@@ -309,6 +322,30 @@ func (n *KongController) toDeckContent(
 	}
 
 	return &content, nil
+}
+func getFCertificateFromKongCert(kongCert kong.Certificate) file.FCertificate {
+	var res file.FCertificate
+	if kongCert.ID != nil {
+		res.ID = kong.String(*kongCert.ID)
+	}
+	if kongCert.Key != nil {
+		res.Key = kong.String(*kongCert.Key)
+	}
+	if kongCert.Cert != nil {
+		res.Cert = kong.String(*kongCert.Cert)
+	}
+	res.SNIs = getSNIs(kongCert.SNIs)
+	return res
+}
+
+func getSNIs(names []*string) []kong.SNI {
+	var snis []kong.SNI
+	for _, name := range names {
+		snis = append(snis, kong.SNI{
+			Name: kong.String(*name),
+		})
+	}
+	return snis
 }
 
 func pluginString(plugin file.FPlugin) string {