+++ internal/ingress/controller/parser/parser_test.go
@@ -6,6 +6,7 @@ import (
 	"sort"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/hbagdi/go-kong/kong"
 	configurationv1 "github.com/kong/kubernetes-ingress-controller/internal/apis/configuration/v1"
@@ -1398,11 +1399,17 @@ func TestParserSecret(t *testing.T) {
 			},
 		}
 
+		t1, _ := time.Parse(time.RFC3339, "2006-01-02T15:04:05Z")
+		t2, _ := time.Parse(time.RFC3339, "2006-01-02T15:05:05Z")
 		secrets := []*corev1.Secret{
 			{
 				ObjectMeta: metav1.ObjectMeta{
+					UID:       "3e8edeca-7d23-4e02-84c9-437d11b746a6",
 					Name:      "secret1",
 					Namespace: "default",
+					CreationTimestamp: metav1.Time{
+						Time: t1,
+					},
 				},
 				Data: map[string][]byte{
 					"tls.crt": []byte(tlsPairs[0].Cert),
@@ -1411,8 +1418,12 @@ func TestParserSecret(t *testing.T) {
 			},
 			{
 				ObjectMeta: metav1.ObjectMeta{
+					UID:       "fc28a22c-41e1-4cd6-9099-fd7756ffe58e",
 					Name:      "secret2",
 					Namespace: "ns1",
+					CreationTimestamp: metav1.Time{
+						Time: t2,
+					},
 				},
 				Data: map[string][]byte{
 					"tls.crt": []byte(tlsPairs[0].Cert),
@@ -1431,6 +1442,19 @@ func TestParserSecret(t *testing.T) {
 		assert.NotNil(state)
 		assert.Equal(1, len(state.Certificates),
 			"certificates are de-duplicated")
+
+		sort.SliceStable(state.Certificates[0].SNIs, func(i, j int) bool {
+			return strings.Compare(*state.Certificates[0].SNIs[i],
+				*state.Certificates[0].SNIs[j]) > 0
+		})
+		assert.Equal(Certificate{
+			Certificate: kong.Certificate{
+				ID:   kong.String("3e8edeca-7d23-4e02-84c9-437d11b746a6"),
+				Cert: kong.String(tlsPairs[0].Cert),
+				Key:  kong.String(tlsPairs[0].Key),
+				SNIs: kong.StringSlice("foo.com", "bar.com"),
+			},
+		}, state.Certificates[0])
 	})
 	t.Run("duplicate SNIs", func(t *testing.T) {
 		ingresses := []*networking.Ingress{